{{ cookiecutter.project_slug }}/requirements.txt
bcrypt==3.2.0
black==21.5b2
certifi==2021.5.30
devtools==0.6.1
django==3.2.4
email-validator==1.1.3
fastapi==0.66.0
gunicorn==20.1.0
isort==5.8.0
jinja2==3.0.1
mypy==0.812
passlib==1.7.4
pip-chill==1.0.1
psycopg2==2.8.6
pyhumps==3.0.2
uvicorn==0.14.0
